Extract AppBar widget in HomePage to its own class.
